rework towards standard reader/writer api
