+++ setup.py
@@ -36,6 +36,7 @@ setup(name='telnetlib3',
                'bin/telnet-talker', ],
       author_email='contact@jeffquast.com',
       platforms='any',
+      zip_safe=True,
       keywords=', '.join(('telnet', 'server', 'client', 'bbs', 'mud', 'utf8',
                           'cp437', 'api', 'library', 'asyncio', 'talker',
                           'tulip', )),
