+++ telnetlib3/__init__.py
@@ -1,16 +1,5 @@
-"""
-telnetlib3: a Telnet Protocol implemented in python.
-
-Requires Python 3.3 or 3.4. and the 'asyncio' module
-(to be distributed with python3.4). See the README file
-for details.
-"""
-__author__ = "Jeffrey Quast"
-__url__ = u'https://github.com/jquast/telnetlib3/'
-__copyright__ = "Copyright 2013"
-__credits__ = ["Jim Storch", "Wijnand Modderman-Lenstra"]
-__license__ = 'ISC'
-
+"""telnetlib3: an asyncio Telnet Protocol implemented in python."""
+# pylint: disable=wildcard-import,undefined-variable
 from .server import *
 from .client import *
 from .telsh import *
@@ -24,3 +13,9 @@ __all__ = (server.__all__ +
            telopt.__all__ +
            slc.__all__ +
            conio.__all__)
+
+__author__ = "Jeff Quast"
+__url__ = u'https://github.com/jquast/telnetlib3/'
+__copyright__ = "Copyright 2013"
+__credits__ = ["Jim Storch", "Wijnand Modderman-Lenstra"]
+__license__ = 'ISC'
