+++ version.json
@@ -1,1 +1,1 @@
-{"version": "0.6.0dev"}
+{"version": "1.0.0"}
